add Cacher as depo subcommand
